Created rigging utilities
