+++ gt/tools/auto_rigger/rig_constants.py
@@ -58,8 +58,7 @@ class RiggerConstants:
     LOC_RADIUS_MULTIPLIER_DRIVEN = .8
     LOC_RADIUS_MULTIPLIER_FK = .3
     LOC_RADIUS_MULTIPLIER_IK = .6
-    # Misc
-    ENUM_ROTATE_ORDER = 'xyz:yzx:zxy:xzy:yxz:zyx'
+    LOC_RADIUS_MULTIPLIER_DATA_QUERY = .1
 
 
 class RiggerDriverTypes:
